Slight change in dojo/request API. refs #15182 !strict
dojo/request calls now return a promise that resolves to the handled
data (similar to the old IO calls). The promise also has a "response"
property; this new new property is a promise that resolves to the object
representing the response and can be used to get information like status
and headers (where applicable).